+++ src/pages/api/models.ts
@@ -10,7 +10,6 @@ import {
   WebLLMProvider,
 } from '~/types/LLMProvider'
 import { getOllamaModels } from '~/utils/modelProviders/ollama'
-import { getOpenAIModels } from '~/utils/modelProviders/openai'
 import { getAzureModels } from '~/utils/modelProviders/azure'
 import { getAnthropicModels } from '~/utils/modelProviders/anthropic'
 import { getWebLLMModels, webLLMModels } from '~/utils/modelProviders/WebLLM'
@@ -18,6 +17,7 @@ import { NextRequest, NextResponse } from 'next/server'
 import { kv } from '@vercel/kv'
 import { getNCSAHostedModels } from '~/utils/modelProviders/NCSAHosted'
 import { migrateAllKeys } from './UIUC-api/MIGRATEALLKEYS'
+import { getOpenAIModels } from '~/app/api/chat/openAI/route'
 
 export const config = {
   runtime: 'edge',
